+++ templates/workspace-template/src/domain/index.ts
@@ -1,0 +1,3 @@
+export * from "./domain.types";
+export * from "./domain.service";
+